requirements.txt
asgiref==3.3.4
Django==3.2.6
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.2
pathlib==1.0.1
boto3==1.18.29
phonenumbers==8.12.22
django-storages==1.11.1
django-filter==2.4.0
django-cors-headers==3.7.0
django-guardian==2.4.0
django-phonenumber-field==5.1.0
django-allauth==0.45.0
dj-rest-auth==2.1.11
djangorestframework-simplejwt==4.8.0
django-elasticsearch-dsl==7.2.0
django-ses==2.2.1
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
drf_yasg==1.20.0
# Need for image processing/managing photo uploads, etc
Pillow==8.3.2

# Prod Only Requirements
gunicorn==20.1.0
psycopg2==2.8.6

# Development/Data Engineering Requirements
django-extensions==3.1.3
ipython

# Testing only requirements
pytest==6.2.4
pytest-django==4.4.0
pytest-mock>=3.6.1
pytest-lazy-fixture==0.6.3
Faker>=8.11.0
coverage>=6.0.1

# Code-Formatting (Local only but sometimes might be useful for remote)
black>=21.7b0
isort>=5.9.3

# Data Load Requirements
git+https://github.com/erikriver/opengraph.git@e2322563004c923a4c1ce136733a44efe5fc8caa
furl==2.1.2
# 3P OpenGraph API used as a fallback (Paid API)
opengraphio==1.0.5

